feat: create filter to customer

+++ src/models/Customers.js
@@ -4,7 +4,7 @@ import { sequelizeConnection } from '../database/db.js'
 const Customers = sequelizeConnection.define('customers', {
   id: { type: DataTypes.INTEGER, primaryKey: true, autoIncrement: true, allowNull: false },
   name: { type: DataTypes.STRING, allowNull: false },
-  number: { type: DataTypes.INTEGER, allowNull: false },
+  number: { type: DataTypes.BIGINT, allowNull: false },
   state: { type: DataTypes.INTEGER, allowNull: false, defaultValue: 1 }
 })
